requirements.txt
appdirs==1.4.3
astroid==1.4.9
click==6.7
Flask==0.12
Flask-PyMongo==0.4.1
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.5
lazy-object-proxy==1.2.2
MarkupSafe==1.0
mccabe==0.6.1
packaging==16.8
pylint==1.6.5
pymongo==3.4.0
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
Werkzeug==0.12.1
wrapt==1.10.10
